Handle and generate stream error within fileUpload.js.
