+++ lib/Error.js
@@ -73,13 +73,3 @@ _Error.StripeRateLimitError = StripeError.extend({type: 'StripeRateLimitError'})
 _Error.StripeConnectionError = StripeError.extend({type: 'StripeConnectionError'});
 _Error.StripeSignatureVerificationError = StripeError.extend({type: 'StripeSignatureVerificationError'});
 _Error.StripeIdempotencyError = StripeError.extend({type: 'StripeIdempotencyError'});
-
-// Error while processing stream for file upload:
-_Error.StreamProcessingError = _Error.extend({
-  type: 'StreamProcessingError',
-  populate: function(raw) {
-    this.type = this.type;
-    this.message = raw.message;
-    this.detail = raw.detail;
-  }
-});
