+++ lib/StripeResource.js
@@ -212,20 +212,6 @@ StripeResource.prototype = {
     }
   },
 
-  _errorHandlerStream: function(callback) {
-    var self = this;
-    return function(error) {
-      callback.call(
-        self,
-        new Error.StreamProcessingError({
-          message: 'An error occurred while attempting to process the file for upload.',
-          detail: error,
-        }),
-        null
-      );
-    }
-  },
-
   _defaultHeaders: function(auth, contentLength, apiVersion) {
     var userAgentString = 'Stripe/v1 NodeBindings/' + this._stripe.getConstant('PACKAGE_VERSION');
 
@@ -260,8 +246,7 @@ StripeResource.prototype = {
       var headers;
 
       if (error) {
-        var handleError = self._errorHandlerStream(callback);
-        return handleError(error);
+        return callback(error);
       }
 
       apiVersion = self._stripe.getApiField('version');
