+++ lib/resources/FileUploads.js
@@ -5,6 +5,7 @@ var utils = require('../utils');
 var StripeResource = require('../StripeResource');
 var stripeMethod = StripeResource.method;
 var multipartDataGenerator = require('../MultipartDataGenerator');
+var Error = require('../Error');
 
 module.exports = StripeResource.extend({
 
@@ -39,8 +40,29 @@ module.exports = StripeResource.extend({
         var buffer = fn(method, bufferData, headers);
         callback(null, buffer);
       }).on('error', function(err) {
-        callback(err, null);
+        var errorHandler = streamError(callback);
+        errorHandler(err, null);
       });
+    }
+
+    function streamError(callback) {
+      var StreamProcessingError = Error.extend({
+        type: 'StreamProcessingError',
+        populate: function(raw) {
+          this.type = this.type;
+          this.message = raw.message;
+          this.detail = raw.detail;
+        }
+      });
+      return function(error) {
+        callback(
+          new StreamProcessingError({
+            message: 'An error occurred while attempting to process the file for upload.',
+            detail: error,
+          }),
+          null
+        );
+      }
     }
   },
